extract page helper
